Move validate_transaction_event_collector_json into validator directory
