+++ tests/agent_features/test_collector_payloads.py
@@ -17,7 +17,6 @@ import webtest
 from testing_support.fixtures import (
     override_application_settings,
     validate_custom_event_collector_json,
-    validate_transaction_event_collector_json,
 )
 from testing_support.sample_applications import (
     simple_app,
@@ -32,6 +31,9 @@ from testing_support.validators.validate_error_trace_collector_json import (
 )
 from testing_support.validators.validate_log_event_collector_json import (
     validate_log_event_collector_json,
+)
+from testing_support.validators.validate_transaction_event_collector_json import (
+    validate_transaction_event_collector_json,
 )
 from testing_support.validators.validate_tt_collector_json import (
     validate_tt_collector_json,
